requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
Django==3.2.3
django-bootstrap3==15.0.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-login-required-middleware==0.6.1
django-tables2==2.4.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tabulate==0.8.9
uWSGI==2.0.19.1
xlrd==2.0.1
xlwt==1.3.0
